Rename kernel_utils to compute_kernel_common to align naming with other algorithms
